Build Update, master branch (2024.12.23.)

See merge request atomki-nuclear-phys/cda!15

+++ devices/v785/Device.cxx
@@ -50,7 +50,7 @@ namespace v785 {
          std::unique_ptr< ChannelConfig > channel( new ChannelConfig() );
          CHECK( channel->readConfig( dev ) );
          if( ( channel->getChannel() >= 0 ) &&
-             ( channel->getChannel() < NUMBER_OF_CHANNELS ) ) {
+             ( channel->getChannel() < static_cast<int>(NUMBER_OF_CHANNELS) ) ) {
             if( m_channels[ channel->getChannel() ] ) {
                m_logger << msg::WARNING
                         << tr( "Redefining channel number: %1" )
@@ -84,7 +84,7 @@ namespace v785 {
 
       // Count the number of configured channels:
       quint32 number_of_channels = 0;
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) ++number_of_channels;
       }
 
@@ -92,7 +92,7 @@ namespace v785 {
       output << number_of_channels;
 
       // Write the channel configurations:
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) {
             CHECK( m_channels[ i ]->writeConfig( dev ) );
          }
@@ -151,7 +151,7 @@ namespace v785 {
 
          // Make sure that the reading was successful:
          if( ( channel->getChannel() >= 0 ) &&
-             ( channel->getChannel() < NUMBER_OF_CHANNELS ) ) {
+             ( channel->getChannel() < static_cast<int>(NUMBER_OF_CHANNELS) ) ) {
             if( m_channels[ channel->getChannel() ] ) {
                m_logger << msg::WARNING
                         << tr( "Redefining channel number: %1" )
@@ -185,7 +185,7 @@ namespace v785 {
       element.setAttribute( "ValidSuppression", m_validSuppressionEnabled );
 
       // Create a new node for the configuration of each channel:
-      for( int i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
+      for( size_t i = 0; i < NUMBER_OF_CHANNELS; ++i ) {
          if( m_channels[ i ] ) {
             QDomElement ch_element =
                element.ownerDocument().createElement( "Channel" );
